+++ plugin/ArkEventHunt/src/plugin_version.h
@@ -1,2 +1,2 @@
 #pragma once
-#define ARKLAND_PLUGIN_VERSION "0.5.2"
+#define ARKLAND_PLUGIN_VERSION "0.5.3"